refactor: teams analytics API 호출 로직 일부 수정
