--- yaxim/src/main/java/com/yaxim/dashboard/exception/GraphApiCallFailedException.java
@@ -1,10 +1,0 @@
-package com.yaxim.dashboard.exception;
-
-import com.yaxim.global.error.model.CustomException;
-import com.yaxim.global.error.model.ErrorCode;
-
-public class GraphApiCallFailedException extends CustomException {
-    public GraphApiCallFailedException() {
-        super(ErrorCode.GRAPH_API_CALL_FAILED);
-    }
-}
