+++ yaxim/src/main/java/com/yaxim/global/error/model/ErrorCode.java
@@ -39,11 +39,15 @@ public enum ErrorCode {
     TEAM_NOT_FOUND(HttpStatus.NOT_FOUND, "TEAM_NOT_FOUND" ),
     OIDC_TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "OIDC_TOKEN_EXPIRED" ),
     TEAM_MEMBER_NOT_MAPPED(HttpStatus.NOT_FOUND, "TEAM_MEMBER_NOT_MAPPED"),
+    USER_ACTIVITY_COUNTS_API_CALL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "USER_ACTIVITY_COUNTS_API_CALL_FAILED" ),
+    USER_ACTIVITY_USER_DETAIL_API_CALL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "USER_ACTIVITY_USER_DETAIL_API_CALL_FAILED" ),
+    TEAM_ACTIVITY_DETAIL_API_CALL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "TEAM_ACTIVITY_DETAIL_API_CALL_FAILED" ),
+    TEAM_ACTIVITY_COUNTS_API_CALL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "TEAM_ACTIVITY_COUNTS_API_CALL_FAILED" ),
     TEAMS_ANALYTICS_NOT_FOUND(HttpStatus.NOT_FOUND, "TEAMS_ANALYTICS_NOT_FOUND"),
-    GRAPH_API_CALL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "GRAPH_API_CALL_FAILED"),
   
     GIT_INFO_NOT_FOUND(HttpStatus.NOT_FOUND, "GIT_INFO_NOT_FOUND" ),
-    GIT_EMAIL_IS_NOT_PROVIDED(HttpStatus.BAD_REQUEST, "GIT_EMAIL_IS_NOT_PROVIDED" ),;
+    GIT_EMAIL_IS_NOT_PROVIDED(HttpStatus.BAD_REQUEST, "GIT_EMAIL_IS_NOT_PROVIDED" ),
+    ;
 
     private final HttpStatus status;
     private final String message;
