+++ yaxim/src/main/java/com/yaxim/graph/controller/dto/PersonalDashboardResponse.java
@@ -1,4 +1,4 @@
-package com.yaxim.dashboard.controller.dto.response;
+package com.yaxim.graph.controller.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -12,7 +12,7 @@ public class PersonalDashboardResponse {
     private final PersonalMetrics personalMetrics;
 
     // 📈 일주일 활동 데이터 (차트용)
-    private final List<PersonalWeeklyData> weeklyChart;
+//    private final List<PersonalWeeklyData> weeklyChart;
 
     @Getter
     @Builder
@@ -25,16 +25,16 @@ public class PersonalDashboardResponse {
         private final Integer urgentMessages;       // 긴급 메시지
         private final Integer totalActivity;        // 총 활동
     }
-
-    @Getter
-    @Builder
-    public static class PersonalWeeklyData {
-        private final String date;                  // "06-01"
-        private final Integer teamChats;            // 팀채팅
-        private final Integer privateChats;         // 개인채팅
-        private final Integer posts;                // 게시물
-        private final Integer meetings;             // 미팅
-        private final Integer calls;                // 통화
-        private final Integer dailyTotal;           // 일일 총합
-    }
+//
+//    @Getter
+//    @Builder
+//    public static class PersonalWeeklyData {
+//        private final String date;                  // "06-01"
+//        private final Integer teamChats;            // 팀채팅
+//        private final Integer privateChats;         // 개인채팅
+//        private final Integer posts;                // 게시물
+//        private final Integer meetings;             // 미팅
+//        private final Integer calls;                // 통화
+//        private final Integer dailyTotal;           // 일일 총합
+//    }
 }
