+++ yaxim/src/main/java/com/yaxim/graph/exception/TeamActivityDetailApiCallFailed.java
@@ -1,0 +1,10 @@
+package com.yaxim.graph.exception;
+
+import com.yaxim.global.error.model.CustomException;
+import com.yaxim.global.error.model.ErrorCode;
+
+public class TeamActivityDetailApiCallFailed extends CustomException {
+    public TeamActivityDetailApiCallFailed() {
+        super(ErrorCode.TEAM_ACTIVITY_DETAIL_API_CALL_FAILED);
+    }
+}
