+++ yaxim/src/main/java/com/yaxim/team/entity/Team.java
@@ -1,6 +1,6 @@
 package com.yaxim.team.entity;
 
-import com.yaxim.global.graph.GraphTeamMemberResponse;
+import com.yaxim.graph.controller.dto.GraphTeamMemberResponse;
 import com.yaxim.global.util.BaseEntity;
 import com.yaxim.team.repository.TeamMemberRepository;
 import com.yaxim.user.entity.UserRole;
