+++ yaxim/src/main/resources/messages.properties
@@ -30,6 +30,11 @@ TOKEN_NOT_PROVIDED=토큰 정보가 존재하지 않습니다.
 TEAM_NOT_FOUND=팀을 찾을 수 없습니다.
 OIDC_TOKEN_EXPIRED=OIDC 토큰이 만료되었습니다.
 TEAM_MEMBER_NOT_MAPPED=해당 이메일로 등록된 팀 구성원이 없습니다.
+USER_ACTIVITY_COUNTS_API_CALL_FAILED=유저 활동량 분석 API 호출에 실패했습니다.
+USER_ACTIVITY_USER_DETAIL_API_CALL_FAILED=유저 활동 상세 분석 API 호출에 실패했습니다.
+TEAM_ACTIVITY_DETAIL_API_CALL_FAILED=팀 활동 상세 분석 API 호출에 실패했습니다.
+TEAM_ACTIVITY_COUNTS_API_CALL_FAILED=팀 활동량 분석 API 호출에 실패했습니다.
+TEAMS_ANALYTICS_NOT_FOUND=팀 분석 내용을 찾을 수 없습니다.
 
 GIT_INFO_NOT_FOUND=깃 정보가 존재하지 않습니다.
 GIT_EMAIL_IS_NOT_PROVIDED=깃 이메일을 설정해주세요.
