+++ yaxim/src/main/resources/messages_en_US.properties
@@ -30,6 +30,11 @@ TOKEN_NOT_PROVIDED=Token Is Not Provided
 TEAM_NOT_FOUND=Team Not Found
 OIDC_TOKEN_EXPIRED=OIDC Token Has Expired
 TEAM_MEMBER_NOT_MAPPED=Team Member Not Mapped By This Email
+USER_ACTIVITY_COUNTS_API_CALL_FAILED=User Activity Count API Call Failed
+USER_ACTIVITY_USER_DETAIL_API_CALL_FAILED=User Activity User Detail API Call Failed
+TEAM_ACTIVITY_DETAIL_API_CALL_FAILED=Team Activity Detail API Call Failed
+TEAM_ACTIVITY_COUNTS_API_CALL_FAILED=Team Activity Counts API Cal Failed
+TEAMS_ANALYTICS_NOT_FOUND=Teams Analytics Not Found
 
 GIT_INFO_NOT_FOUND=Git Info Not Found
 GIT_EMAIL_IS_NOT_PROVIDED=Git Email Is Not Provided
